+++ src/browser/js/app/widgets/plots/led.js
@@ -9,6 +9,7 @@ module.exports = class Led extends Widget {
         return {
             type:'led',
             id:'auto',
+            linkId:'',
 
             _geometry:'geometry',
 
@@ -25,7 +26,6 @@ module.exports = class Led extends Widget {
 
             _led:'led',
 
-            widgetId:'',
             range:{min:0,max:1},
             logScale:false,
             value:'',
@@ -41,6 +41,9 @@ module.exports = class Led extends Widget {
 
     constructor(options) {
 
+        // backward compat
+        if (options.props.widgetId) options.props.value = '@{' + options.props.widgetId + '}'
+
         var html = `
             <div class="led">
             </div>
@@ -48,28 +51,14 @@ module.exports = class Led extends Widget {
 
         super({...options, html: html})
 
-        if (this.getProp('widgetId').length) widgetManager.on(`change.${this.hash}`, this.syncHandle.bind(this))
-
         this.setValue(this.getProp('range').min)
 
     }
 
-    syncHandle(e) {
-
-        if (this.getProp('widgetId')!=e.id || !widgetManager.getWidgetById(e.id).length) return
-        var widget = widgetManager.getWidgetById(e.id),
-            value
-        for (var i=widget.length-1; i>=0; i--) {
-            if (widget[i].getValue) {
-                this.setValue(widget[i].getValue(), {sync: e.options.sync})
-                return
-            }
-        }
-
-    }
     setValue(v, options={}) {
 
         if (typeof v != 'number') return
+        
         this.value = v
         this.widget.style.setProperty('--opacity', mapToScale(v,[this.getProp('range').min,this.getProp('range').max],[0,1],false,this.getProp('logScale'),true))